fix warning

+++ src/PubSub/PubSubBuild.php
@@ -5,7 +5,9 @@ class PubSubBuild {
 	public function build ($root) {
 		$cache = [];
 		$listersCache = $root . '/../subscribers/_build.php';
-		unlink($listersCache);
+		if (file_exists($listersCache)) {
+			unlink($listersCache);
+		}
 		$files = glob($root . '/../subscribers/*.php');
 		if (count($files) == 0) {
 			file_put_contents($listersCache, '<?php' . "\n" . 'return [];');
